0923時点作業記録
