+++ build.gradle
@@ -21,8 +21,6 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-cache'
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 	
-// 	implementation 'org.springframework.boot:spring-boot-starter-jdbc'
-
 	implementation 'org.springframework.boot:spring-boot-starter-data-jdbc'
 	implementation 'org.springframework.boot:spring-boot-starter-security'
 	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
@@ -34,19 +32,14 @@ dependencies {
 
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	runtimeOnly 'com.h2database:h2'
+
+//	implementation 'com.github.springtestdbunit:spring-test-dbunit:1.3.0'
+//	implementation 'org.dbunit:dbunit:2.5.1'
+
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-//  	testCompile group: 'org.dbunit', name: 'dbunit', version: '2.6.0'
-
-	// 以下を参考にライブラリを追加する
-	// https://qiita.com/a-pompom/items/3f834119c756e5286730#%E5%AF%BE%E8%B1%A1%E8%AA%AD%E8%80%85%E3%82%B4%E3%83%BC%E3%83%AB
-//	testCompile group: 'com.github.springtestdbunit', name: 'spring-test-dbunit', version: '1.3.0' // SpringでDBUnitを用いる際に必要
-//	testCompile group: 'org.dbunit', name: 'dbunit', version: '2.5.1' 
-	implementation 'com.github.springtestdbunit:spring-test-dbunit:1.3.0'
-	implementation 'org.dbunit:dbunit:2.5.1'
-	implementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter-test:1.3.4'
-	
+	testImplementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter-test:2.1.3'
 	testImplementation 'org.springframework.security:spring-security-test'
+
 	compileOnly("org.projectlombok:lombok")
 }
 
@@ -55,7 +48,6 @@ dependencyManagement {
 		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
 	}
 }
-
 
 test {
 	useJUnitPlatform()
