+++ src/main/java/com/example/demo/DemoApplication.java
@@ -3,6 +3,8 @@ package com.example.demo;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import com.example.demo.repository.SampleRepository;
+
 //@EntityScan("com.example.demo.repository")
 @SpringBootApplication
 public class DemoApplication {
@@ -10,5 +12,11 @@ public class DemoApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(DemoApplication.class, args);
 	}
+	
+	   private final SampleRepository todoMapper;
 
+	    public DemoApplication(SampleRepository todoMapper) {
+	        this.todoMapper = todoMapper; // Mapperをインジェクションする
+	    }
+	    
 }
