+++ src/main/resources/data.sql
@@ -1,1 +1,2 @@
 INSERT INTO todo ( todo_id, todo_title, finished, created_at ) VALUES ( 1, '朝ごはんつくる', false, '2020-12-01 09:15:00' );
+INSERT INTO todo ( todo_id, todo_title, finished, created_at ) VALUES ( 2, '夜ごはんつくる', false, '2020-12-01 09:15:00' );
