--- src/test/java/com/example/demo/CsvDataSetLoader.java
@@ -1,16 +1,0 @@
-package com.example.demo;
-
-import org.dbunit.dataset.IDataSet;
-import org.dbunit.dataset.csv.CsvURLDataSet;
-import org.springframework.core.io.Resource;
-
-import com.github.springtestdbunit.dataset.AbstractDataSetLoader;
-
-public class CsvDataSetLoader extends AbstractDataSetLoader {
-
-	@Override
-	protected IDataSet createDataSet(Resource resource) throws Exception {
-		return new CsvURLDataSet(resource.getURL());
-	}
-
-}
